+++ src/card/sets/y2021/modern_horizons_2.rs
@@ -1033,9 +1033,10 @@ pub(in crate::card::sets) static GRIST_THE_HUNGER_TIDE: CardRecord = CardRecord:
                 "\u{2212}2: You may sacrifice a creature.",
                 &[CostDef::Loyalty(-2)],
                 EffectDef::PayOr(PayOrDef::optional(
-                    &[CostDef::SacrificePermanentMatching(
-                        ObjectPredicateDef::HasType(CardType::Creature),
-                    )],
+                    &[CostDef::SacrificePermanent {
+                        object: ObjectPredicateDef::HasType(CardType::Creature),
+                        controller: PlayerRelation::You,
+                    }],
                     &EffectDef::None,
                 )),
             ),